Hopefully made the server possible for Linux Change exceptionCaught Fixed the bugs that other people had, because I don't have them??
git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@112 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/net/server/handlers/channel/AranComboHandler.java
@@ -35,10 +35,10 @@ public class AranComboHandler extends AbstractMaplePacketHandler {
         MapleCharacter player = c.getPlayer();
         if (player.getJobType() == 2) { //Keep it this till Evan comes in Private Servers.
             if (player.getCombo() > 0 && (System.currentTimeMillis() - player.getLastAttack() > 3000)) {
-                player.setCombo(0);
+                player.setCombo((short) 0);
                 player.cancelBuffStats(MapleBuffStat.ARAN_COMBO);
             } else {
-                int combo = player.getCombo() + 1;
+                short combo = (short) (player.getCombo() + 1);
 	    switch (combo) {
 		case 10:
 		case 20:
