+++ src/net/server/handlers/channel/RangedAttackHandler.java
@@ -56,13 +56,13 @@ public final class RangedAttackHandler extends AbstractDealDamageHandler {
         } else if (attack.skill == Aran.COMBO_SMASH || attack.skill == Aran.COMBO_PENRIL || attack.skill == Aran.COMBO_TEMPEST) {
             player.getMap().broadcastMessage(player, MaplePacketCreator.rangedAttack(player, attack.skill, attack.skilllevel, attack.stance, attack.numAttackedAndDamage, 0, attack.allDamage, attack.speed, attack.direction, attack.display), false);            
             if (attack.skill == Aran.COMBO_SMASH && player.getCombo() >= 30) {
-                player.setCombo(0);
+                player.setCombo((short) 0);
                 applyAttack(attack, player, 1);
             } else if (attack.skill == Aran.COMBO_PENRIL && player.getCombo() >= 100) {
-                player.setCombo(0);
+                player.setCombo((short) 0);
                 applyAttack(attack, player, 2);
             } else if (attack.skill == Aran.COMBO_TEMPEST && player.getCombo() >= 200) {
-                player.setCombo(0);
+                player.setCombo((short) 0);
                 applyAttack(attack, player, 4);    
             }
         } else {
